To be continued.....

+++ Webzine.WebApplication/Areas/administration/Controllers/StyleController.cs
@@ -1,7 +1,6 @@
 ﻿namespace Webzine.WebApplication.Areas.Admin.Controllers
 {
     using Microsoft.AspNetCore.Mvc;
-    using Webzine.Entity;
     using Microsoft.EntityFrameworkCore;
     using Webzine.Repository.Contracts;
     using Webzine.WebApplication.Areas.administration.ViewModels;
@@ -10,14 +9,10 @@
     /// Représente le controlleur pour la partie des Styles dans la zone d'administration.
     /// </summary>
     [Area("administration")]
-    [Route("[controller]")]
     public class StyleController : Controller
     {
         private IStyleRepository _styleRepository;
         private StyleViewModel model = new StyleViewModel();
-
-        [BindProperty]
-        public Style style { get; set; }
 
         public StyleController(IStyleRepository styleRepository)
         {
@@ -25,23 +20,20 @@
         }
 
         /// <summary>
-        /// Page par défaut de la partie administration de Style
+        /// Page par défaut de la partie administration de Style.
         /// </summary>
         /// <returns></returns>
-        [Route("")]
-        [HttpGet("[action]")]
         public IActionResult Index()
         {
-            var style = this._styleRepository.FindAll().ToList();
-            this.model.Styles = style;
+            var style = 
+            this.model.Styles = this._styleRepository.FindAll().ToList();
             return this.View(this.model);
         }
 
         /// <summary>
-        /// Page de création d'un Style  
+        /// Page de création d'un Style.
         /// </summary>
         /// <returns></returns>
-        [HttpGet("[action]")]
         public IActionResult Create()
         {
             return this.View();
@@ -55,7 +47,6 @@
         [HttpPost]
         [ValidateAntiForgeryToken]
         [ActionName("create")]
-        [Route("[action]")]
         public async Task<IActionResult> Create(StyleViewModel model)
         {
             try
@@ -63,17 +54,16 @@
                 if (this.ModelState.IsValid)
                 {
                     this._styleRepository.Add(model.Style);
-                    var tot = model.Style;
-                    return RedirectToAction(nameof(Index));
+
+                    return this.RedirectToAction(nameof(this.Index));
                 }
             }
             catch (DbUpdateException /* ex */)
             {
-                //Log the error (uncomment ex variable name and write a log.
-                ModelState.AddModelError("", "Unable to save changes. " +
-                    "Try again, and if the problem persists " +
-                    "see your system administrator.");
+                // Log the error (uncomment ex variable name and write a log.
+                ModelState.AddModelError(string.Empty, "Unable to save changes. " + "Try again, and if the problem persists " +"see your system administrator.");
             }
+
             return this.View(model.Style);
         }
 
@@ -82,11 +72,9 @@
         /// </summary>
         /// <param name="idStyle"></param>
         /// <returns></returns>
-        [HttpGet("[action]")]
         public IActionResult Edit(int id)
         {
-            var style = this._styleRepository.Find(id);
-            this.model.Style = style;
+            this.model.Style = this._styleRepository.Find(id);
             return this.View(this.model);
         }
 
@@ -98,7 +86,6 @@
         [HttpPost]
         [ValidateAntiForgeryToken]
         [ActionName("edit")]
-        [Route("[action]")]
         public async Task<IActionResult> Edit(StyleViewModel model, int id)
         {
             try
@@ -107,42 +94,44 @@
                 {
                     model.Style.IdStyle = id;
                     this._styleRepository.Update(model.Style);
-                    var tot = model.Style;
-                    return RedirectToAction(nameof(Index));
+                    return this.RedirectToAction(nameof(this.Index));
                 }
             }
             catch (DbUpdateException /* ex */)
             {
-                //Log the error (uncomment ex variable name and write a log.
-                ModelState.AddModelError("", "Unable to save changes. " +
-                    "Try again, and if the problem persists " +
-                    "see your system administrator.");
+                // Log the error (uncomment ex variable name and write a log.
+                this.ModelState.AddModelError(string.Empty, "Unable to save changes. " +"Try again, and if the problem persists " +"see your system administrator.");
             }
+
             return this.View(model.Style);
         }
 
         /// <summary>
-        /// Page de suppression du Style passé en paramètre
+        /// Page de suppression du Style passé en paramètre.
         /// </summary>
         /// <param name="idStyle"></param>
         /// <returns></returns>
-        [HttpGet("[action]")]
         public IActionResult Delete(int id)
         {
-            var style = this._styleRepository.Find(id);
-            this.model.Style = style;
+            this.model.Style = this._styleRepository.Find(id);
             return this.View(this.model);
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         [ActionName("delete")]
-        [Route("[action]")]
         public IActionResult Delete(StyleViewModel model, int id)
         {
-            Style style = this._styleRepository.Find(id);
-            this._styleRepository.Delete(style);
-            return this.RedirectToAction("Index");
+            var style = this._styleRepository.Find(id);
+            if (this.ModelState.IsValid)
+            {
+                this._styleRepository.Delete(style);
+                return this.RedirectToAction("Index");
+            }
+            else
+            {
+                return this.View(style);
+            }
         }
     }
 }
